fix(core): fixed readonly bug during patch apply
Trrack stores diffs between state as patches, and the issue arises when applying the patches. Say if
patch A contains a complex object, and patch B changes one of the keys, when trrack tries to apply
patch B after patch A it runs into readonly error. This happens because patches are readonly.